requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
passlib[bcrypt]
python-jose[cryptography]
fastapi-mail
python-multipart
python-dotenv
redis
geopy
pyjwt
httpx
pydantic-settings
